btrfs-progs: add encrypt property support
BTRFS way is to set the properties using the xattr functions,
and now to continue on that this patch adds a way to add the
encryption property to the subvolume.

The property name that is newly being introduced here is
'btrfs.encrypt' or just 'encrypt'. In future this property
may take more number of encryption mode. For now the supported
current mode with this patch is 'fscryptv1' which instructs
to use fs/crypto V1 encryption mode/method.

Signed-off-by: Anand Jain <anand.jain@oracle.com>