btrfs-progs: split global help HELPINFO_INSERT_GLOBALS
As of now the define HELPINFO_INSERT_GLOBALS if used as in the example
as below (as of now its not been used anywhere) will print the help
texts as shown below

#diff --git a/cmds/filesystem.c b/cmds/filesystem.c
#index 4f22089abeaa..564dc40cc99a 100644
#--- a/cmds/filesystem.c
#+++ b/cmds/filesystem.c
#@@ -631,6 +631,7 @@ static const char * const
#cmd_filesystem_show_usage[] = {
#        "-m|--mounted       show only mounted btrfs",
#        HELPINFO_UNITS_LONG,
#        "If no argument is given, structure of all present filesystems
#is shown.",
#+       HELPINFO_INSERT_GLOBALS,
#        NULL
# };
#
$ ./btrfs fi show --help

 <snip>

    Global options:
    --format TYPE      where TYPE is: text

$

So in preparation to add --verbose and --quiet global options, and
apparently --format is not being used yet, this patch splits the global
options into two defines.

                                       "Global options:"

So that the currently added global options --verbose and --quiet can use
the define HELPINFO_INSERT_GLOBALS header as shown below.

(For example:)
$ ./btrfs fi show --help
<snip>

    Global options:

Signed-off-by: Anand Jain <anand.jain@oracle.com>