btrfs-progs: update device add ioctl with device type
When a device is added to the mounted device provide an optonal suffix
to add device roles similar to mkfs.btrfs;

  btrfs device add /dev/sda[:m|:d|:monly|:dataonly]

btrfs-progs: update device add ioctl to support device roles

The btrfs device add command now accepts an optional suffix to specify
device roles, mirroring the functionality introduced in mkfs.btrfs. This
allows users to define the role of a device (metadata preferred, data
preferred, metadata only, or data only) at the time of adding it to a
mounted filesystem.

Example:

  btrfs device add /dev/sdb:d /mnt/btrfs

	This command adds /dev/sdb to the mounted filesystem at /mnt/btrfs
	and designates its role as preferred for data. The supported
	suffixes are :m, :d, :monly, and :donly.

Signed-off-by: Anand Jain <anand.jain@oracle.com>