btrfs-progs: disable exclusive metadata/data device roles
When exclusive device roles are set, the actual number of devices
available for allocation depends on the block group type (metadata or
data), rather than the total number of devices in the filesystem. The
total number of devices is widely used in the code; for example, `device
remove` and `mount` have to validate if they would satisfy the block group
profile based on the device roles.

These necessary changes will be implemented in both the kernel,
btrfs-progs, and fstests in a subsequent update, once the current
metadata-preferred and data-preferred device roles are stable. For now,
the metadata-only and data-only roles are marked as unsupported.

However, devices with metadata-preferred and data-preferred roles are
included in the total device count. Therefore, the above changes are not
required to support these roles, and the preferred roles are ready.

Signed-off-by: Anand Jain <anand.jain@oracle.com>